move fix to patch3d so every artiests can use it
